pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llmwiki"
version = "0.1.0"
description = "LLM-powered personal knowledge base — raw data goes in, an LLM compiles it into a structured, interlinked wiki"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [{name = "Huang Geyang", email = "Sukebeta@outlook.com"}]
keywords = ["llm", "knowledge-base", "wiki", "mcp", "karpathy"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Text Processing :: Markup :: Markdown",
]
dependencies = [
    "openai>=1.30.0",
    "python-dotenv>=1.0.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
    "flask>=3.0.0",
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0",
    "markdownify>=0.13.0",
    "matplotlib>=3.8.0",
    "python-frontmatter>=1.1.0",
    "watchdog>=4.0.0",
    "pymupdf>=1.24.0",
    "httpx>=0.27.0",
    "mcp>=1.0.0",
]

[project.optional-dependencies]
cjk = ["opencc-python-reimplemented>=0.1.7"]
all = ["llmbase[cjk]"]

[project.urls]
Homepage = "https://github.com/Hosuke/llmbase"
Documentation = "https://github.com/Hosuke/llmbase/tree/main/docs"
Repository = "https://github.com/Hosuke/llmbase"
Demo = "https://huazangge-production.up.railway.app"

[project.scripts]
llmbase = "tools.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.setuptools.packages.find]
include = ["tools*"]
